pdf2txt/requirements.txt
torch
torchvision
timm==0.5.4
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
scipy
opencv-python
pdf2image
easyocr
os
argparse
regex
symspellpy
time
'git+https://github.com/facebookresearch/detectron2.git'
